Updated the service implementation structure
