+++ src/main/java/com/icops/DaraManagement/service/UserService.java
@@ -3,8 +3,6 @@ package com.icops.DaraManagement.service;
 
 import com.icops.DaraManagement.model.User;
 import com.icops.DaraManagement.model.entity.UserModel;
-import com.icops.DaraManagement.repository.UserDao;
-import org.springframework.beans.factory.annotation.Autowired;
 
 public interface UserService {
