+++ src/main/java/com/icops/DaraManagement/service/impl/ParentServiceImpl.java
@@ -1,10 +1,13 @@
 package com.icops.DaraManagement.service.impl;
 
 import com.icops.DaraManagement.model.Parent;
+import com.icops.DaraManagement.model.enums.Gender;
 import com.icops.DaraManagement.repository.ParentDao;
 import com.icops.DaraManagement.service.ParentService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class ParentServiceImpl implements ParentService {
@@ -13,5 +16,34 @@ public class ParentServiceImpl implements ParentService {
 
     public Parent create(final Parent parent) {
         return parentDao.save(parent);
+    }
+
+    public List<Parent> allParents() {
+        return parentDao.findAll();
+    }
+
+    public Parent findById(Long id) {
+        Parent parent = parentDao.findById(id).orElse(null);
+        return parent;
+    }
+
+    public List<Parent> findByStudent(Long id) {
+        return  parentDao.findByStudent(id);
+    }
+
+    public Long findStudentIdByLastname(String lastName){
+        Long st= parentDao.findStudentIdByLastname(lastName);
+        return st;
+    }
+    public List<Parent> findByGender(Gender gender) {return parentDao.findByGender(gender);}
+    public Parent updateParent(Parent parent, Long id)
+    {
+
+        Parent updatedParent = parentDao.save(parent);
+        return updatedParent;
+    }
+
+    public void deleteParent(Long id) {
+        parentDao.deleteById(id);
     }
 }
